feat: improved external queue Add method and test cases (#17)
* feat: included job count method for group job

* refactor: update queue Add methods to return success boolean and merge PQItem into Item struct

* test: add comprehensive test cases for queue resize functionality

* fix: correct queue enqueue logic by moving postEnqueue after successful enqueue

* fix: add error handling for JSON marshaling in persistent queues

* refactor: rename JobCount to Len and extract atomic counter into groupLen type

* fix: correct atomic decrement operation in groupLen Sub method

* test: add queue tests and move Len() to EnqueuedGroupJob interface

* test: fix test assertions and reorder worker start in queue tests

* fix: move wait group increment from constructor to NewJob method for accurate group job tracking

* fix: move waitgroup increment from NewJob to constructor for proper concurrency handling

* test: improve queue test assertions with value verification and refactor duplicated constants

+++ group_job_test.go
@@ -10,7 +10,7 @@ import (
 func TestGroupJob(t *testing.T) {
 	t.Run("group job creation with newGroupJob", func(t *testing.T) {
 		// Create a new group job with buffer size 3
-		bufferSize := uint32(3)
+		bufferSize := 3
 		gj := newGroupJob[string, int](bufferSize)
 
 		assert := assert.New(t)
@@ -36,7 +36,7 @@ func TestGroupJob(t *testing.T) {
 
 	t.Run("creating new job in a group", func(t *testing.T) {
 		// Create a new group job
-		bufferSize := uint32(3)
+		bufferSize := 3
 		gj := newGroupJob[string, int](bufferSize)
 
 		// Create a job within the group
@@ -56,12 +56,12 @@ func TestGroupJob(t *testing.T) {
 
 	t.Run("closing a group job", func(t *testing.T) {
 		// Create a group job with small buffer to avoid wait group blocking
-		bufferSize := uint32(1)
+		bufferSize := 1
 		gj := newGroupJob[string, int](bufferSize)
 
 		// Manually reduce the wait group count to match our single Close() call
 		// This prevents the wait group from blocking indefinitely
-		gj.wg.Add(-int(bufferSize) + 1)
+		gj.wg.Add(-bufferSize + 1)
 
 		assert := assert.New(t)
 
@@ -79,7 +79,7 @@ func TestGroupJob(t *testing.T) {
 
 	t.Run("job status transitions in a group job", func(t *testing.T) {
 		// Create a new group job
-		bufferSize := uint32(1)
+		bufferSize := 1
 		gj := newGroupJob[string, int](bufferSize)
 
 		// Reduce wait group count to prevent deadlock when testing
